test(core): split grid legend tests
