+++ packages/core/src/view/gridView/gridView.test.js
@@ -9,8 +9,6 @@ import ViewRenderingContext from "../renderingContext/viewRenderingContext.js";
 import { normalizeClipOptions } from "../renderingContext/clipOptions.js";
 import UnitView from "../unitView.js";
 import AxisView from "../axisView.js";
-import LegendView from "../legendView.js";
-import { translateLegendCoords } from "./gridView.js";
 import { createAndInitialize, createTestViewContext } from "../testUtils.js";
 import { createLogicalVisibleRect } from "../../marks/mark.js";
 
@@ -80,18 +78,6 @@ class InspectRenderingContext extends ViewRenderingContext {
                 normalizeClipOptions(options)
             ),
         });
-    }
-}
-
-class MarkRecordingRenderingContext extends NoOpRenderingContext {
-    /** @type {string[]} */
-    markNames = [];
-
-    /**
-     * @param {import("../../marks/mark.js").default} mark
-     */
-    renderMark(mark) {
-        this.markNames.push(mark.unitView.name);
     }
 }
 
@@ -169,420 +155,6 @@ const recordRenderOrder = (spec, labels) => {
         return order;
     });
 };
-
-describe("translateLegendCoords", () => {
-    test("places a right-oriented legend next to the viewport", () => {
-        const legendView = /** @type {any} */ ({
-            getPerpendicularSize: () => 80,
-            getExternalPadding: () => 12,
-        });
-
-        const coords = translateLegendCoords(
-            Rectangle.create(10, 20, 300, 200),
-            "right",
-            legendView
-        );
-
-        expect(coords.x).toBe(322);
-        expect(coords.y).toBe(20);
-        expect(coords.width).toBe(80);
-        expect(coords.height).toBe(200);
-    });
-});
-
-describe("GridView legends", () => {
-    const createLegendTestView = (
-        /** @type {Partial<import("../../spec/root.js").RootSpec>} */ spec = {}
-    ) =>
-        createAndInitialize(
-            /** @type {import("../../spec/root.js").RootSpec} */ ({
-                vconcat: [
-                    {
-                        data: {
-                            values: [
-                                { x: 1, y: 2, Origin: "Europe" },
-                                { x: 2, y: 3, Origin: "Japan" },
-                            ],
-                        },
-                        mark: "point",
-                        encoding: {
-                            x: { field: "x", type: "quantitative" },
-                            y: { field: "y", type: "quantitative" },
-                            color: { field: "Origin", type: "nominal" },
-                        },
-                    },
-                ],
-                ...spec,
-            }),
-            ConcatView
-        );
-
-    const getLegends = (/** @type {ConcatView} */ view) =>
-        view
-            .getDescendants()
-            .filter((descendant) => descendant instanceof LegendView);
-
-    test("keeps legends hidden by default", async () => {
-        const view = await createLegendTestView();
-
-        expect(getLegends(view)).toHaveLength(0);
-    });
-
-    test("creates an opt-in right legend for a nominal color scale", async () => {
-        const view = await createLegendTestView({
-            config: { legend: { disable: false } },
-        });
-        const legends = getLegends(view);
-        const labels = legends[0]
-            .getDescendants()
-            .find((descendant) => descendant.name == "labels");
-
-        expect(legends).toHaveLength(1);
-        expect(legends[0].name).toBe("legend_right");
-        expect(legends[0].legendProps.title).toBe("Origin");
-        expect(labels).toBeInstanceOf(UnitView);
-        expect(
-            Array.from(labels.flowHandle.collector.getData()).map(
-                ({ value, label, _legendIndex }) => ({
-                    value,
-                    label,
-                    _legendIndex,
-                })
-            )
-        ).toEqual([
-            { value: "Europe", label: "Europe", _legendIndex: 0 },
-            { value: "Japan", label: "Japan", _legendIndex: 1 },
-        ]);
-    });
-
-    test("merges redundant color and shape channels into one symbol legend", async () => {
-        const view = await createLegendTestView({
-            config: { legend: { disable: false } },
-            vconcat: [
-                {
-                    data: {
-                        values: [
-                            { x: 1, y: 2, Origin: "Europe" },
-                            { x: 2, y: 3, Origin: "Japan" },
-                        ],
-                    },
-                    mark: "point",
-                    encoding: {
-                        x: { field: "x", type: "quantitative" },
-                        y: { field: "y", type: "quantitative" },
-                        color: { field: "Origin", type: "nominal" },
-                        shape: { field: "Origin", type: "nominal" },
-                    },
-                },
-            ],
-        });
-        const legends = getLegends(view);
-
-        expect(legends).toHaveLength(1);
-    });
-
-    test("creates an explicit channel legend even when defaults are disabled", async () => {
-        const view = await createLegendTestView({
-            vconcat: [
-                {
-                    data: {
-                        values: [
-                            { x: 1, y: 2, Origin: "Europe" },
-                            { x: 2, y: 3, Origin: "Japan" },
-                        ],
-                    },
-                    mark: "point",
-                    encoding: {
-                        x: { field: "x", type: "quantitative" },
-                        y: { field: "y", type: "quantitative" },
-                        color: {
-                            field: "Origin",
-                            type: "nominal",
-                            legend: { title: "Region" },
-                        },
-                    },
-                },
-            ],
-        });
-        const legends = getLegends(view);
-
-        expect(legends).toHaveLength(1);
-        expect(legends[0].legendProps.title).toBe("Region");
-    });
-
-    test("respects explicit legend null", async () => {
-        const view = await createLegendTestView({
-            config: { legend: { disable: false } },
-            vconcat: [
-                {
-                    data: {
-                        values: [
-                            { x: 1, y: 2, Origin: "Europe" },
-                            { x: 2, y: 3, Origin: "Japan" },
-                        ],
-                    },
-                    mark: "point",
-                    encoding: {
-                        x: { field: "x", type: "quantitative" },
-                        y: { field: "y", type: "quantitative" },
-                        color: {
-                            field: "Origin",
-                            type: "nominal",
-                            legend: null,
-                        },
-                    },
-                },
-            ],
-        });
-
-        expect(getLegends(view)).toHaveLength(0);
-    });
-
-    test("does not create legends for positional quantitative channels", async () => {
-        const view = await createLegendTestView({
-            config: { legend: { disable: false } },
-            vconcat: [
-                {
-                    data: { values: [{ x: 1, y: 2 }] },
-                    mark: "point",
-                    encoding: {
-                        x: { field: "x", type: "quantitative" },
-                        y: { field: "y", type: "quantitative" },
-                    },
-                },
-            ],
-        });
-
-        expect(getLegends(view)).toHaveLength(0);
-    });
-
-    test("creates an opt-in gradient legend for quantitative color", async () => {
-        const view = await createLegendTestView({
-            config: { legend: { disable: false } },
-            vconcat: [
-                {
-                    data: {
-                        values: [
-                            { x: 1, y: 2, measurement: 0 },
-                            { x: 2, y: 3, measurement: 1 },
-                        ],
-                    },
-                    mark: "rect",
-                    encoding: {
-                        x: { field: "x", type: "index" },
-                        y: { field: "y", type: "index" },
-                        color: {
-                            field: "measurement",
-                            type: "quantitative",
-                        },
-                    },
-                },
-            ],
-        });
-        const legends = getLegends(view);
-        const ramp = legends[0]
-            .getDescendants()
-            .find((descendant) => descendant.name == "gradientRamp");
-        const labels = legends[0]
-            .getDescendants()
-            .find((descendant) => descendant.name == "gradientLabels");
-        const plot = view
-            .getDescendants()
-            .find((descendant) => descendant.name == "grid0");
-
-        expect(legends).toHaveLength(1);
-        expect(ramp).toBeInstanceOf(UnitView);
-        expect(labels).toBeInstanceOf(UnitView);
-        expect(plot).toBeInstanceOf(UnitView);
-        expect(
-            /** @type {UnitView} */ (ramp).getScaleResolution("y").getScale()
-                .props
-        ).toEqual(expect.objectContaining({ domainTransition: false }));
-        expect(/** @type {UnitView} */ (ramp).getScaleResolution("color")).toBe(
-            /** @type {UnitView} */ (plot).getScaleResolution("color")
-        );
-        const rampData = Array.from(
-            /** @type {UnitView} */ (ramp).flowHandle.collector.getData()
-        );
-        const labelData = Array.from(
-            /** @type {UnitView} */ (labels).flowHandle.collector.getData()
-        );
-
-        expect(rampData.length).toBeGreaterThan(1);
-        expect(rampData[0]).toEqual(
-            expect.objectContaining({
-                position0: 0,
-                value: expect.any(Number),
-            })
-        );
-        expect(rampData.at(-1)).toEqual(
-            expect.objectContaining({
-                position1: 1,
-                value: expect.any(Number),
-            })
-        );
-        expect(labelData.length).toBeGreaterThan(1);
-        expect(
-            labelData.every(({ position }) => position >= 0 && position <= 1)
-        ).toBe(true);
-        expect(labelData.every(({ label }) => typeof label == "string")).toBe(
-            true
-        );
-
-        const context = new MarkRecordingRenderingContext({ picking: false });
-        view.render(context, Rectangle.create(0, 0, 700, 300), {
-            firstFacet: true,
-        });
-        expect(context.markNames).toEqual(
-            expect.arrayContaining([
-                "gradientRamp",
-                "gradientTicks",
-                "gradientLabels",
-            ])
-        );
-    });
-
-    test("gradient legends mirror source scale type and color scheme", async () => {
-        const view = await createLegendTestView({
-            config: { legend: { disable: false } },
-            vconcat: [
-                {
-                    data: {
-                        values: [
-                            { x: 1, y: 1, measurement: 1 },
-                            { x: 2, y: 2, measurement: 100 },
-                        ],
-                    },
-                    mark: "point",
-                    encoding: {
-                        x: { field: "x", type: "quantitative" },
-                        y: { field: "y", type: "quantitative" },
-                        color: {
-                            field: "measurement",
-                            type: "quantitative",
-                            scale: {
-                                type: "log",
-                                domain: [1, 100],
-                                scheme: "turbo",
-                            },
-                        },
-                    },
-                },
-            ],
-        });
-        const legend = getLegends(view)[0];
-        const ramp = legend
-            .getDescendants()
-            .find((descendant) => descendant.name == "gradientRamp");
-        const labels = legend
-            .getDescendants()
-            .find((descendant) => descendant.name == "gradientLabels");
-        const plot = view
-            .getDescendants()
-            .find((descendant) => descendant.name == "grid0");
-
-        expect(ramp).toBeInstanceOf(UnitView);
-        expect(labels).toBeInstanceOf(UnitView);
-        expect(plot).toBeInstanceOf(UnitView);
-        expect(/** @type {UnitView} */ (ramp).getScaleResolution("color")).toBe(
-            /** @type {UnitView} */ (plot).getScaleResolution("color")
-        );
-
-        const labelData = Array.from(
-            /** @type {UnitView} */ (labels).flowHandle.collector.getData()
-        );
-        expect(labelData).toEqual(
-            expect.arrayContaining([
-                expect.objectContaining({ value: 1, position: 0 }),
-                expect.objectContaining({ value: 10, position: 0.5 }),
-                expect.objectContaining({ value: 100, position: 1 }),
-            ])
-        );
-    });
-
-    test("threshold gradient legends include outer color buckets", async () => {
-        const view = await createLegendTestView({
-            config: { legend: { disable: false } },
-            vconcat: [
-                {
-                    data: {
-                        values: [
-                            { x: 1, y: 1, measurement: 10 },
-                            { x: 2, y: 2, measurement: 110 },
-                        ],
-                    },
-                    mark: "point",
-                    encoding: {
-                        x: { field: "x", type: "quantitative" },
-                        y: { field: "y", type: "quantitative" },
-                        color: {
-                            field: "measurement",
-                            type: "quantitative",
-                            scale: {
-                                type: "threshold",
-                                domain: [20, 40, 60, 80, 100],
-                                scheme: "turbo",
-                            },
-                        },
-                    },
-                },
-            ],
-        });
-        const legend = getLegends(view)[0];
-        const ramp = legend
-            .getDescendants()
-            .find((descendant) => descendant.name == "gradientRamp");
-        const labels = legend
-            .getDescendants()
-            .find((descendant) => descendant.name == "gradientLabels");
-
-        expect(ramp).toBeInstanceOf(UnitView);
-        expect(labels).toBeInstanceOf(UnitView);
-
-        const rampData = Array.from(
-            /** @type {UnitView} */ (ramp).flowHandle.collector.getData()
-        );
-        const labelData = Array.from(
-            /** @type {UnitView} */ (labels).flowHandle.collector.getData()
-        );
-
-        expect(rampData).toHaveLength(6);
-        expect(rampData[0].position0).toBe(0);
-        expect(rampData[0].value).toBeLessThan(20);
-        expect(rampData.at(-1).position1).toBe(1);
-        expect(rampData.at(-1).value).toBeGreaterThan(100);
-        expect(labelData[0]).toEqual(
-            expect.objectContaining({
-                value: 20,
-                position: rampData[0].position1,
-            })
-        );
-        expect(labelData.at(-1)).toEqual(
-            expect.objectContaining({
-                value: 100,
-                position: rampData.at(-2).position1,
-            })
-        );
-    });
-
-    test("does not draw configured view strokes inside legends", async () => {
-        const view = await createLegendTestView({
-            config: {
-                legend: { disable: false },
-                view: { stroke: "lightgray" },
-            },
-        });
-        const legends = getLegends(view);
-        const legendBackgroundStrokes = legends[0]
-            .getDescendants()
-            .filter((descendant) =>
-                descendant.name.startsWith("backgroundStroke")
-            );
-
-        expect(legends).toHaveLength(1);
-        expect(legendBackgroundStrokes).toHaveLength(0);
-    });
-});
 
 describe("GridView incremental child management", () => {
     test("dynamic add/remove keeps shared axes in sync", async () => {
